+++ front/src/app/App.tsx
@@ -10,10 +10,12 @@ import { Toaster } from '~/components/ui/toaster';
 import CalendarPage from '~/features/calendar/CalendarPage';
 import FollowingPage from '~/features/following/FollowingPage';
 import Footer from '~/features/footer/Footer';
-import Header from '~/features/header/Header';
+import DesktopHeader from '~/features/header/DesktopHeader';
+import MobileHeader from '~/features/header/MobileHeader';
 import SearchPage from '~/features/search/SearchPage';
 import SettingsPage from '~/features/settings/SettingsPage';
 import ShowPage from '~/features/show/ShowPage';
+import { useIsMobile } from '~/hooks/useIsMobile';
 import { useAppDispatch, useAppSelector } from '~/store';
 import { fetchfollowedShowsAction } from '~/store/user/actions';
 import { selectIsLoggedIn } from '~/store/user/selectors';
@@ -25,6 +27,7 @@ import { ROUTES } from './routes';
 const App = () => {
   const dispatch = useAppDispatch();
   const isLoggedIn = useAppSelector(selectIsLoggedIn);
+  const isMobile = useIsMobile();
 
   useEffect(() => {
     if (isLoggedIn) {
@@ -43,7 +46,7 @@ const App = () => {
       <Toaster />
 
       <Flex direction="column" minH="97vh">
-        <Header />
+        {isMobile ? <MobileHeader /> : <DesktopHeader />}
 
         <ErrorBoundary>
           <Switch>
