+++ front/src/features/header/Logo.tsx
@@ -1,0 +1,47 @@
+import { Button, Flex, Image } from '@chakra-ui/react';
+import { useLocation } from 'wouter';
+
+import { ROUTES } from '~/app/routes';
+import logo from '~/assets/images/logo.svg';
+import { setShouldResetSearchInput } from '~/features/search/searchInputSlice';
+import { useIsMobile } from '~/hooks/useIsMobile';
+import { useNavigateWithAnimation } from '~/hooks/useNavigateWithAnimation';
+import { useAppDispatch } from '~/store';
+import { applyViewTransition } from '~/utils/applyViewTransition';
+
+interface LogoProps {
+  onClose?: () => void;
+}
+
+const Logo = ({ onClose }: LogoProps) => {
+  const isMobile = useIsMobile();
+  const dispatch = useAppDispatch();
+  const [location] = useLocation();
+  const navigate = useNavigateWithAnimation();
+
+  const handleLogoClick = () => {
+    onClose?.();
+    if (location === ROUTES.HOME) {
+      applyViewTransition(() => dispatch(setShouldResetSearchInput(true)));
+    } else {
+      navigate(ROUTES.HOME);
+    }
+  };
+
+  return (
+    <Flex align="center" as="h1" m={{ base: '0 auto', md: 'unset' }}>
+      <Button variant="plain" onClick={handleLogoClick}>
+        <Image
+          alt="TV Minder logo"
+          display="inline"
+          h="30px"
+          mt={isMobile ? '' : '6px'}
+          src={logo}
+          verticalAlign="middle"
+        />
+      </Button>
+    </Flex>
+  );
+};
+
+export default Logo;
